Add serverCloser and specDeleter structs.
The former struct encapsulates a reference to a (concurrently serving) http.Server
and a channel used to receive the error produced by http.Server#Serve.
Its Close() method shuts down the server and returns the produced error.

The latter extends the former to by additionally deleting the socker
that was listened to.